Finished refactor moved test classes back to original location.
