+++ api/src/apiTest/java/com/ford/labs/retroquest/api/TeamApiTest.java
@@ -1,6 +1,6 @@
-package com.ford.labs.retroquest.api.refactor;
+package com.ford.labs.retroquest.api;
 
-import com.ford.labs.retroquest.api.refactor.setup.ApiTest;
+import com.ford.labs.retroquest.api.setup.ApiTest;
 import com.ford.labs.retroquest.team.*;
 import org.junit.After;
 import org.junit.Test;
